Refactor service structure
