+++ src/api/services/github/config.ts
@@ -1,0 +1,7 @@
+import { GitHubConfig } from "./types";
+
+export const getGitHubConfig = (): GitHubConfig => ({
+  cacheTTL: 5 * 60, // 5 minutes in seconds
+  circuitBreakerFailures: 3,
+  circuitBreakerResetTime: 30000, // 30 seconds
+});
