+++ src/api/services/redis/types.ts
@@ -1,0 +1,14 @@
+export interface RedisConfig {
+  url: string;
+  retryStrategy: {
+    maxDelay: number;
+    initialDelay: number;
+  };
+  maxRetriesPerRequest: number;
+}
+
+export interface RedisError extends Error {
+  code?: string;
+  command?: string;
+  args?: string[];
+}
